chore: migrate jsii-pacmak tests to jest
