+++ packages/jsii-pacmak/package.json
@@ -12,7 +12,7 @@
     "build": "npm run gen && tsc --build && chmod +x bin/jsii-pacmak && tslint -p .",
     "watch": "tsc --build -w",
     "lint": "tslint -p . --force",
-    "test": "/bin/bash test/diff-test.sh && /bin/bash test/build-test.sh && nodeunit test/test.*.js",
+    "test": "/bin/bash test/diff-test.sh && /bin/bash test/build-test.sh && jest",
     "package": "package-js"
   },
   "keywords": [
@@ -37,18 +37,34 @@
     "@types/commonmark": "^0.27.4",
     "@types/escape-string-regexp": "^2.0.1",
     "@types/fs-extra": "^8.0.0",
+    "@types/jest": "^24.0.18",
     "@types/node": "^8.10.54",
-    "@types/nodeunit": "^0.0.30",
     "@types/xmlbuilder": "^11.0.1",
     "@types/yargs": "^13.0.2",
+    "jest": "^24.9.0",
     "jsii-build-tools": "file:../jsii-build-tools",
     "jsii-calc": "file:../jsii-calc",
     "jsii-dotnet-jsonmodel": "file:../jsii-dotnet-jsonmodel",
     "jsii-dotnet-runtime": "file:../jsii-dotnet-runtime",
     "jsii-java-runtime": "file:../jsii-java-runtime",
-    "nodeunit": "^0.11.3",
     "tslint": "^5.20.0",
     "typescript": "~3.6.3"
+  },
+  "jest": {
+    "collectCoverage": true,
+    "collectCoverageFrom": [
+      "**/bin/**/*.js",
+      "**/lib/**/*.js"
+    ],
+    "coverageReporters": [
+      "lcov",
+      "text"
+    ],
+    "errorOnDeprecated": true,
+    "testEnvironment": "node",
+    "testMatch": [
+      "**/?(*.)+(spec|test).js"
+    ]
   },
   "author": {
     "name": "Amazon Web Services",
